Refactor advection diffusion namespace
Move the vorticity-only AdvectionDiffusion class into the
spatiotemporal namespace and point the advection_diffusion config at
it directly. Remove the old top-level module and unused helper solver
code so stable imports come from one hierarchy.
