+++ docs/_toc.yml
@@ -22,12 +22,12 @@ chapters:
           - file: reference/simulations/base
           - file: reference/simulations/spatiotemporal/index
             sections:
+              - file: reference/simulations/spatiotemporal/advection_diffusion
               - file: reference/simulations/spatiotemporal/advection_diffusion_multichannel
               - file: reference/simulations/spatiotemporal/gray_scott
               - file: reference/simulations/spatiotemporal/gross_pitaevskii
               - file: reference/simulations/spatiotemporal/navier_stokes_conditioned
               - file: reference/simulations/spatiotemporal/reaction_diffusion
-          - file: reference/simulations/advection_diffusion
           - file: reference/simulations/double_pendulum
           - file: reference/simulations/epidemic
           - file: reference/simulations/flow_problem
