--- docs/reference/simulations/advection_diffusion.rst
@@ -1,12 +1,0 @@
-autosim.simulations.advection_diffusion
-=======================================
-
-.. deprecated:: 0.1.0
-
-   Use ``autosim.simulations.spatiotemporal.AdvectionDiffusionMultichannel`` with
-   ``output_indices=[0]`` for vorticity-only advection-diffusion data.
-
-.. automodule:: autosim.simulations.advection_diffusion
-   :members:
-   :undoc-members:
-   :show-inheritance:
