+++ docs/reference/simulations/spatiotemporal/advection_diffusion.rst
@@ -1,0 +1,7 @@
+autosim.simulations.spatiotemporal.advection_diffusion
+======================================================================
+
+.. automodule:: autosim.simulations.spatiotemporal.advection_diffusion
+   :members:
+   :undoc-members:
+   :show-inheritance:
