+++ src/autosim/simulations/spatiotemporal/__init__.py
@@ -1,5 +1,6 @@
 """Spatiotemporal simulator implementations."""
 
+from .advection_diffusion import AdvectionDiffusion
 from .advection_diffusion_multichannel import AdvectionDiffusionMultichannel
 from .gray_scott import GrayScott
 from .gross_pitaevskii import GrossPitaevskiiEquation2D
@@ -7,6 +8,7 @@ from .navier_stokes_conditioned import ConditionedNavierStokes2D
 from .reaction_diffusion import ReactionDiffusion
 
 __all__ = [
+    "AdvectionDiffusion",
     "AdvectionDiffusionMultichannel",
     "ConditionedNavierStokes2D",
     "GrayScott",
