+++ tests/test_cli.py
@@ -44,8 +44,7 @@ class DummySimulator(SpatioTemporalSimulator):
 def test_build_simulator_from_target_core_and_experimental() -> None:
     core_cfg = OmegaConf.create(
         {
-            "_target_": "autosim.simulations.spatiotemporal."
-            "AdvectionDiffusionMultichannel",
+            "_target_": "autosim.simulations.spatiotemporal.AdvectionDiffusion",
             "log_level": "warning",
         }
     )
@@ -56,9 +55,7 @@ def test_build_simulator_from_target_core_and_experimental() -> None:
         }
     )
 
-    assert (
-        build_simulator(core_cfg).__class__.__name__ == "AdvectionDiffusionMultichannel"
-    )
+    assert build_simulator(core_cfg).__class__.__name__ == "AdvectionDiffusion"
     assert build_simulator(experimental_cfg).__class__.__name__ == "ShallowWater2D"
